% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.bacon_summary}
\alias{plot.bacon_summary}
\title{Plot a Bacon Decomposition}
\usage{
\method{plot}{bacon_summary}(x, ...)
}
\arguments{
\item{x}{A \code{bacon_summary} object returned by \code{\link[=bacon_weight_summary]{bacon_weight_summary()}}.}

\item{...}{Additional arguments passed to \code{\link[=autoplot.bacon_summary]{autoplot.bacon_summary()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Thin wrapper around \code{\link[=autoplot.bacon_summary]{autoplot.bacon_summary()}} that prints the scatter plot
to the current graphics device.
}
