% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_diagnostics_report.R
\name{print.did_report}
\alias{print.did_report}
\title{Print a DiD Diagnostics Report}
\usage{
\method{print}{did_report}(x, ...)
}
\arguments{
\item{x}{A \code{did_report} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a DiD Diagnostics Report
}
